requirements.txt
streamlit
google-genai
plotly
pandas
python-dotenv
streamlit-extras
streamlit-option-menu
streamlit-antd-components